HL7 messages are not shown in web admin: Error: java.lang.Long cannot be cast to java.lang.Integer | Request method 'GET' not supported - TRUNK-3576 Author: ulakkaraju
git-svn-id: http://svn.openmrs.org/openmrs/branches/1.9.x@28589 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/hl7/db/hibernate/HibernateHL7DAO.java
@@ -35,6 +35,7 @@ import org.openmrs.hl7.HL7InQueue;
 import org.openmrs.hl7.HL7Source;
 import org.openmrs.hl7.Hl7InArchivesMigrateThread;
 import org.openmrs.hl7.db.HL7DAO;
+import org.openmrs.util.OpenmrsUtil;
 
 /**
  * OpenMRS HL7 API database default hibernate implementation This class shouldn't be instantiated by
@@ -177,14 +178,14 @@ public class HibernateHL7DAO implements HL7DAO {
 	}
 	
 	/**
-	 * @see org.openmrs.hl7.db.HL7DAO#countHL7s(java.lang.Class, java.lang.Integer,
+	 * @see org.openmrs.hl7.db.HL7DAO#countHL7s(java.lang.Class, java.lang.Long,
 	 *      java.lang.String)
 	 */
 	@SuppressWarnings("rawtypes")
-	public Integer countHL7s(Class clazz, Integer messageState, String query) {
+	public Long countHL7s(Class clazz, Integer messageState, String query) {
 		Criteria crit = getHL7SearchCriteria(clazz, messageState, query);
 		crit.setProjection(Projections.rowCount());
-		return (Integer) crit.uniqueResult();
+		return (Long)crit.uniqueResult();
 	}
 	
 	/**
